refactor(core): split markdown and wikilinks into peer sub-packages
Carves apps/core/markdown.py and apps/core/markdown_links.py into two
sibling sub-packages under apps/core/:

- apps/core/wikilinks/ — type-system primitive. Adds types.py (LinkType
  + renderer registry, extracted from markdown_links.py) alongside the
  existing base.py (WikilinkableModel) and picker.py (PickerType).
- apps/core/markdown/ — consumer. Splits into render.py (markdown→HTML
  pipeline + wikilink rendering), field.py (MarkdownField + conversion,
  no RecordReference), and references.py (the markdown→RecordReference
  bridge: sync_references, save_inline_markdown_field).

MarkdownField and get_markdown_fields move out of apps/core/models.py
to markdown/field.py. Migration-invisible: MarkdownField.deconstruct()
already returns TextField. RecordReference and register_reference_cleanup
stay in models.py as foundational core infrastructure.

Catalog model files import MarkdownField via apps.core.markdown without
dragging in the RecordReference graph: __init__.py loads field.py and
render.py only; references.py is imported directly by save-path callers
so the dependency surface stays explicit.

All importers flipped in one PR (no compat shims). Test files
reorganized: link-registry tests moved to test_wikilinks.py; rendering,
conversion, and RecordReference syncing tests merged into test_markdown.py;
test_markdown_links.py deleted.
