+++ backend/apps/catalog/models/series.py
@@ -6,8 +6,8 @@ from typing import TYPE_CHECKING
 
 from django.db import models
 
+from apps.core.markdown import MarkdownField
 from apps.core.models import (
-    MarkdownField,
     SluggedModel,
     TimeStampedModel,
     field_not_blank,
