+++ backend/apps/catalog/models/taxonomy.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING, ClassVar
 from django.db import models
 from django.db.models.functions import Lower
 
+from apps.core.markdown import MarkdownField
 from apps.core.models import (
-    MarkdownField,
     SluggedModel,
     TimeStampedModel,
     field_not_blank,
