--- backend/apps/core/markdown.py
@@ -1,192 +1,0 @@
-"""Markdown rendering pipeline.
-
-Shared utility for converting markdown text to sanitized HTML. Used by
-API serializers to produce ``{field}_html`` output for markdown fields.
-"""
-
-import re
-from dataclasses import dataclass, field
-from typing import Any
-
-import nh3
-from django.db import models
-from django.utils.safestring import mark_safe
-from markdown_it import MarkdownIt
-
-from apps.core.markdown_links import render_all_links
-from apps.core.models import get_markdown_fields
-
-# CommonMark-compliant markdown parser.
-# - linkify: auto-link bare URLs during parsing (structure-aware, won't
-#   linkify inside code blocks or existing links)
-# - breaks: single newlines become <br> (equiv. to Python-Markdown's nl2br)
-# - typographer + smartquotes/replacements: smart quotes, em dashes, ellipsis
-#   (equiv. to Python-Markdown's smarty)
-# - fenced code blocks are built into the commonmark preset
-_md = MarkdownIt(
-    "commonmark", {"linkify": True, "breaks": True, "typographer": True}
-).enable(["linkify", "replacements", "smartquotes", "table", "strikethrough"])
-
-# Allowed HTML tags for markdown rendering
-ALLOWED_TAGS = {
-    "p",
-    "br",
-    "strong",
-    "em",
-    "s",
-    "ul",
-    "ol",
-    "li",
-    "code",
-    "pre",
-    "blockquote",
-    "a",
-    "h1",
-    "h2",
-    "h3",
-    "h4",
-    "h5",
-    "h6",
-    "img",
-    "hr",
-    "table",
-    "thead",
-    "tbody",
-    "tr",
-    "th",
-    "td",
-    "figure",
-    "figcaption",
-    "sup",
-}
-
-# Allowed attributes per tag
-ALLOWED_ATTRIBUTES = {
-    "a": {"href", "title"},
-    "img": {"src", "alt", "title"},
-    "code": {"class"},
-    "pre": {"class"},
-    "th": {"align"},
-    "td": {"align"},
-    "sup": {"data-cite-id", "data-cite-index", "tabindex", "role"},
-}
-
-# Regex for task list items: matches <li> followed by optional <p>, then [ ], [  ], [], [x], or [X]
-# Group 1: optional whitespace+<p> (for blank-line-separated list items)
-# Group 2: the check character to determine checked state (spaces or empty = unchecked)
-_TASK_LIST_RE = re.compile(r"<li>(\s*<p>)?\s*\[( *|[xX])\]")
-
-
-def _convert_task_list_items(html: str) -> str:
-    """Convert task list markers in <li> tags to checkbox HTML.
-
-    After markdown and nh3 processing, literal [ ] and [x] text inside <li> tags
-    is converted to checkbox inputs. Each checkbox gets a sequential
-    data-checkbox-index attribute for JavaScript targeting.
-
-    This runs AFTER nh3 sanitization, so the injected <input> elements are
-    trusted server code, not user-supplied HTML.
-
-    Args:
-        html: Sanitized HTML containing <li>[ ] or <li>[x] patterns
-
-    Returns:
-        HTML with task list markers replaced by checkbox inputs
-    """
-    counter = 0
-
-    def _replace(match: re.Match) -> str:
-        nonlocal counter
-        idx = counter
-        counter += 1
-        p_tag = (
-            match.group(1) or ""
-        )  # Preserve <p> if present (blank-line-separated items)
-        check_char = match.group(2)
-        checked_attr = " checked" if check_char in ("x", "X") else ""
-        return (
-            f'<li class="task-list-item">{p_tag}'
-            f'<input type="checkbox"{checked_attr} data-checkbox-index="{idx}" disabled>'
-        )
-
-    return _TASK_LIST_RE.sub(_replace, html)
-
-
-# ``metadata_out`` dicts are free-form: each link type's ``collect_metadata``
-# callback defines its own key set and core never inspects them, so a TypedDict
-# would lock in a shape core has no business enforcing.
-def render_markdown_html(
-    text: str | None, metadata_out: list[dict[str, Any]] | None = None
-) -> str:
-    """Convert markdown text to sanitized HTML.
-
-    Full pipeline: wiki links -> markdown (with linkify) -> nh3 -> checkboxes.
-
-    Args:
-        text: Raw markdown text (may contain ``[[type:ref]]`` links).
-        metadata_out: When provided, passed to ``render_all_links`` so
-            link types with ``collect_metadata`` can append structured data.
-
-    Returns:
-        Sanitized HTML ``SafeString``, safe for direct use in templates.
-    """
-    if not text:
-        return ""
-    # Convert [[type:ref]] links to markdown links (before markdown processing)
-    text = render_all_links(text, metadata_out=metadata_out)
-    # Convert markdown to HTML (bare URLs are auto-linked during parsing)
-    html = _md.render(text)
-    # Sanitize to prevent XSS
-    safe_html = nh3.clean(html, tags=ALLOWED_TAGS, attributes=ALLOWED_ATTRIBUTES)
-    # Convert task list markers to checkboxes (after sanitization for security)
-    return mark_safe(_convert_task_list_items(safe_html))  # noqa: S308 — HTML sanitized by nh3
-
-
-@dataclass(frozen=True, slots=True)
-class RenderedField:
-    """Rendered markdown output for a single field.
-
-    ``citations`` stays as ``list[dict]`` because link types are extensible
-    and ``apps.core`` has no business naming provenance-layer schemas (see
-    ``render_markdown_html`` note). Callers validate citations into their
-    own schemas via ``Schema.model_validate(...)`` at the boundary.
-    """
-
-    html: str
-    citations: list[dict[str, Any]] = field(default_factory=list)
-
-
-def render_markdown_field(obj: models.Model, field_name: str) -> RenderedField:
-    """Render one markdown field on *obj* to HTML + collected citations.
-
-    Returns empty ``RenderedField`` if the field's raw value is blank.
-    """
-    citations: list[dict[str, Any]] = []
-    html = render_markdown_html(getattr(obj, field_name, ""), metadata_out=citations)
-    return RenderedField(html=html, citations=citations)
-
-
-# Return value mixes rendered HTML strings and free-form metadata dicts
-# (see ``render_markdown_html`` note); callers spread this into API payloads.
-def render_markdown_fields(
-    obj: models.Model,
-) -> dict[str, str | list[dict[str, Any]]]:
-    """Return ``{field}_html`` rendered values for all MarkdownField instances on *obj*.
-
-    Designed for use with ``**`` spread in API serialization dicts::
-
-        return {
-            "name": obj.name,
-            "description": obj.description,
-            **render_markdown_fields(obj),
-        }
-    """
-    result: dict[str, str | list[dict[str, Any]]] = {}
-    for field_name in get_markdown_fields(type(obj)):
-        citations: list[dict[str, Any]] = []
-        result[f"{field_name}_html"] = render_markdown_html(
-            getattr(obj, field_name, ""), metadata_out=citations
-        )
-        if citations:
-            result[f"{field_name}_citations"] = citations
-    return result
